portal-service-test Move OrganizationFinderTest of the com.liferay.portal.service.persistence package to organizations-test module
